requirements.txt
Django~=4.1
djangorestframework~=3.13.1
django-filter~=22.1